User class detection in akka http
