+++ instrumentation-security/akka-http-2.11_2.4.5/build.gradle
@@ -1,0 +1,48 @@
+apply plugin: 'scala'
+
+isScalaProjectEnabled(project, "scala-2.11")
+
+
+sourceSets.test.scala.srcDir "src/test/java"
+sourceSets.test.java.srcDirs = []
+
+jar {
+    manifest { attributes 'Implementation-Title': 'com.newrelic.instrumentation.security.akka-http-2.11_2.4.5' }
+}
+
+dependencies {
+    implementation(project(":newrelic-security-api"))
+    implementation("com.newrelic.agent.java:newrelic-api:${nrAPIVersion}")
+    implementation("com.newrelic.agent.java:newrelic-weaver-api:${nrAPIVersion}")
+    implementation("com.typesafe.akka:akka-http_2.11:10.1.8")
+    implementation("com.typesafe.akka:akka-stream_2.11:2.5.19")
+    implementation("com.typesafe.akka:akka-actor_2.11:2.5.19")
+}
+
+verifyInstrumentation {
+    passesOnly('com.typesafe.akka:akka-http_2.11:[10.0.0,)') {
+        implementation("com.typesafe.akka:akka-stream_2.11:2.5.19")
+    }
+    passesOnly('com.typesafe.akka:akka-http_2.12:[10.0.0,)') {
+        implementation("com.typesafe.akka:akka-stream_2.11:2.5.19")
+    }
+
+    passesOnly('com.typesafe.akka:akka-http_2.13:[10.1.8,)') {
+        implementation("com.typesafe.akka:akka-stream_2.13:2.5.23")
+        implementation("com.typesafe.akka:akka-http_2.13:10.1.8")
+        implementation("com.typesafe.akka:akka-actor_2.13:2.5.23")
+    }
+
+    excludeRegex 'com.typesafe.akka:akka-http-experimental_2.11:.*$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.11:.*(RC|M)[0-9]*$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.12:.*(RC|M)[0-9]*$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.13:.*(RC|M)[0-9]*$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.11:.*-[0-9a-f]{8}$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.12:.*-[0-9a-f]{8}$'
+    excludeRegex 'com.typesafe.akka:akka-http_2.13:.*-[0-9a-f]{8}$'
+}
+
+site {
+    title 'Akka Http'
+    type 'Framework'
+}
