+++ settings.gradle
@@ -115,7 +115,7 @@ include 'instrumentation:javax-xpath'
 include 'instrumentation:akka-http-core-2.11_10.0.11'
 include 'instrumentation:akka-http-core-2.13_10.2.0'
 include 'instrumentation:akka-http-core-10.0'
-include 'instrumentation:akka-http-2.13_10.1.8'
+include 'instrumentation:akka-http-2.11_2.4.5'
 include 'instrumentation:jetty-9'
 include 'instrumentation:jetty-11'
 //include 'instrumentation:netty-4.0.0'
